Fixes compiling on windows 32-bit (#1549)
* use correct LONG type for nni_atomic_flag on win32

* use InterlockExchangeAdd for nni_atomic_get_bool

- this is equivelent to InterlockAdd for the purposes of this call (since it is adding 0)
- this allows the code to compile on 32bit windows

+++ src/platform/windows/win_impl.h
@@ -56,7 +56,7 @@ struct nni_plat_cv {
 #define NNI_CV_INITIALIZER(mxp) { .srl = mxp, .cv = CONDITION_VARIABLE_INIT }
 
 struct nni_atomic_flag {
-	unsigned f;
+	LONG f;
 };
 
 struct nni_atomic_bool {
